move s3 cache to locaitons_lookup.py
