+++ test/unit/test_main.py
@@ -1,12 +1,11 @@
 from unittest.mock import patch
 import json
 
-from main import init, CACHE, S3Client, parse_params, load_swagger_docs
+from main import parse_params, load_swagger_docs
 from test.unit.test_helpers import TestHelpers
 
 
 class TestMain:
-
     @classmethod
     def setup_class(cls):
         TestHelpers.set_env_vars()
@@ -16,13 +15,6 @@ class TestMain:
     def teardown_class(cls):
         TestHelpers.clear_env_vars()
         TestHelpers.tear_down()
-
-    def test_init_sets_s3_locations(self):
-        with patch.object(S3Client, '__init__', lambda s, b, r: None):
-            with patch.object(S3Client, 'fetch_cache', lambda p: 'locations'):
-                assert CACHE.get('s3_locations') is None
-                init()
-                assert CACHE.get('s3_locations') == 'locations'
 
     def test_parse_params(self):
         params = {'fields': 'url,hours,address', "location_codes": 'abc'}
